Use handlerThread to handle data process
